irbempy: remove double-layer import
